% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autohaplo_graphs.R
\name{autohaplo_graphs}
\alias{autohaplo_graphs}
\title{Generate graphical output of a haplotype analysis}
\usage{
autohaplo_graphs(results, output_dir = ".", graphs = "default", ...)
}
\arguments{
\item{results}{A list generated as output from the function  \code{\link{haplo_selection}}
and thus containing the results of a haplotype analysis.}

\item{output_dir}{A character. The name of the directory to which the files
should be written.}

\item{graphs}{A named list containing at most 6 elements corresponding to the
names of elements in a list output by function \code{\link{haplo_selection}} :
\code{All_markers}, \code{Filtered_markers}, \code{Clustered_markers},
\code{Selected_clusters}, \code{Selected_markers}, \code{Haplotypes}. Each
of the elements of these lists must contain a character vector cotaining at
most 4 elements corresponding to the 4 possible graphs that can be output :
\code{"density"}, \code{"matrix"}, \code{"distance"}, \code{genotypes}. The
list provided will determine which graphs will be output from which objects.}

\item{...}{Arguments passed to function \code{ggplot::ggsave} which writes
the image file to disc.}
}
\value{
\code{NULL}, invisibly. The function is called for its side effect of
  writing graphs to files.
}
\description{
This is a high-level function for taking an object obtained as output from
the function \code{\link{haplo_selection}} and generating graphical output
written to files. The \code{graphs} argument allows a high degree of
coustomization as to the graphs that will be generated. Some graphs input/output
combinations are not allowed (see \code{Details}).
}
\details{
The default \code{graphs} option currently results in no graphs being generated.
Sensible defaults still have to be defined.

The output is directed by default to the current directory. An error is thrown
if the suggested output directory already exists, else a new directory with this
name is created. This is meant to avoid name conflicts and file overwriting,
but this could be changed in the future.

This function calls high-level plotting functions defined in the package itself :
\code{\link{density_plot}}, \code{\link{ld_plot}}, \code{\link{distance_plot}},
\code{\link{genotype_plot}}. These functions are called in the context of
\code{autohaplo_graphs} with their default values, but finer control could be
obtained by using these functions directly.

Arguments can be passed to function \code{ggplot::ggsave}, which directs the
output (\code{width} and \code{height} default to 7). PDF files are generated
by default.
}
\examples{
NULL

}
