% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autohaplo_output.R
\name{autohaplo_output}
\alias{autohaplo_output}
\title{Output the results of a haplotype analysis to disk}
\usage{
autohaplo_output(results, output_dir = NULL, logfile = "Log.txt",
  variant_files = "default",
  variant_format = results$Parameters$File_format, graphs = "default")
}
\arguments{
\item{results}{A list generated as output from the function  \code{\link{haplo_selection}}
and thus containing the results of a haplotype analysis.}

\item{output_dir}{A list generated as output from the function  \code{\link{haplo_selection}}
and thus containing the results of a haplotype analysis.}

\item{logfile}{A character. The name of the file to which the logfile
should be written. Defaults to "Log.txt". If \code{NULL}, no logfile
is written.}

\item{variant_files}{A character vector. The name of the list elements of
\code{results} that should used to output the variant data to disk. Can
be any of \code{All_markers}, \code{Filtered_markers}, \code{Clustered_markers},
\code{Selected_clusters}, \code{Selected_markers}, \code{Haplotypes}.
Defaults to the last three of these elements. If null, no variant file
will be written to disk.}

\item{variant_format}{A character. The format to be used for writing
variant files. Must be either "vcf" or "hapmap". Other formats are not
supported and will result in an error.}

\item{graphs}{A list describing which graphical elements should be output.
See \code{\link{autohaplo_graphs}} for more details.}
}
\value{
\code{NULL}, invisibly. This function is called for its side
  effect of writing output.
}
\description{
This function takes an object returned by function \code{\link{haplo_selection}}
as input and extracts information from this object so as the make the
information easily interpretable by the user. The function directs the
output of essentially three elements : the analysis logfile, the variant
files (vcf or hapmap) containing the marker set at different steps of the
analysis, and various kinds of graphs
}
\details{
More control over the files output by this function can be obtained by
using the underlying functions : \code{\link{autohaplo_logfile}},
\code{\link{autohaplo_graphs}}, \code{\link{write_hapmap}}
}
\examples{
NULL

}
