% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Generate clusters of markers in high linkage disequilibrium}
\usage{
cluster(snp_data, center_pos, block_threshold = 1)
}
\arguments{
\item{snp_data}{A list: the snp data to be clustered. Must contain at least
\code{Genotypes}, \code{LD} and \code{Markers} fields for clustering to
occur properly.}

\item{center_pos}{The central position on either side of which clustering
should be performed.}

\item{block_threshold}{The minimum R^2 for two adjacent markers to be
eligible for blocking (defaults to 1, i.e. perfect linkage disequilibrium).}
}
\value{
A subset of the original \code{snp_data} containing only one
 representative marker per LD block.
}
\description{
This function clusters groups of markers that are in high linkage
disequilibrium on a single side of the central gene position considered.
The objective of the clustering step is to reduce the number of markers
considered and thus allow for a simplification of haplotype definition
problem.
}
\details{
This function is used internally by the package and is not normally
available for end users.
}
\examples{
NULL

}
