% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ld.R
\name{compute_ld}
\alias{compute_ld}
\title{Linkage disequilibrium computation}
\usage{
compute_ld(snp_data, measure, kinship = NA, structure = NA)
}
\arguments{
\item{snp_data}{A list of SNP data comprising at least an element called
\code{Genotypes}, a \code{SnpMatrix} object containing the genotype of
individuals included in the analysis.}

\item{measure}{{A character of length one. The r2 measure that was used
for computing the linkage disequilibrium values in \code{LD}. Can be either
\code{"r2"}, \code{"r2v"}, \code{"r2s"}, \code{"r2vs"}.}}

\item{kinship}{A square numeric matrix of kinship values with as many rows and
columns as there are individuals included in the analysis. More specifically,
\code{kinship} must have as many rows and columns as there are rows in
\code{snp_data$Genotypes}.}

\item{structure}{A numeric matrix with as many rows as there are individuals
in the analysis and as many columns as there are putative subpopulations in
the dataset.}
}
\value{
A list with two components :
  \itemize{
  \item{LD} {A symmetric Matrix of class \code{dsCMatrix} containing linkage
  disequilibrium values for all marker pairs.}
  \item{LD_measure} {A character of length one. The r2 measure that was used
  for computing the linkage disequilibrium values in \code{LD}. Can be either
  \code{"r2"}, \code{"r2v"}, \code{"r2s"}, \code{"r2vs"}.}
  }
}
\description{
This function computes linkage disequilibrium (LD) for a set of selected loci.
The measure can be either r2, r2v, r2s, or r2vs. Function \code{ld} from
package \code{snpStats} is used when computing r2, whereas r2 values corrected
with kinship data (r2v), population structure (r2s), or both (r2vs) make use
of function \code{LD.Measures} from package \code{LDcorSV}.
}
\details{
An error will be thrown if the function is asked to compute an r2 measure
without all required inputs (kinship and/or structure data). If kinship
and/or structure data are given as input but the computed measure makes no
use of these data, they will be ignored with a warning.
}
\examples{
NULL

}
