% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_plot.R
\name{density_plot}
\alias{density_plot}
\title{Plot of marker density along chromosome}
\usage{
density_plot(snp_data, chr_length = NULL, center_pos = NULL)
}
\arguments{
\item{snp_data}{A list comprising at least an element called \code{Markers},
a \code{data.frame} giving information about a set of markers. See
\code{\link{read_hapmap}} and \code{\link{read_vcf}} for more information
about the structure of \code{snp_data}.}

\item{chr_length}{A numeric of length one. The length of the chromosome
(in base pairs). This is used to set the plotting window of the x-axis
from 0 to the length of the chromosome. If \code{NULL} (default), the
plotting window will only cover those markers present in the dataset.}

\item{center_pos}{A numeric of length one. The position of the center of
the gene of interest (in base pairs). This is used to add a red vertical
line at gene position. If \code{NULL}, no vertical line will be added.}
}
\value{
A graphical object inheriting from classes \code{gg} and \code{ggplot}
  from package \code{ggplot2}, which can be either assigned to an object
  or printed to a graphical device.
}
\description{
This is a high-level function to plot marker density along the chromosome.
This may help in identifying regions that are poorly covered by markers and
is thus helpful for troubleshooting.
}
\details{
As the output of this function is a graphical object from package \code{ggplot2},
it is possible to customize the appearance of the graph and add layers to
the graph using the syntax of this package.
}
\examples{
NULL

}
