% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_plot.R
\name{distance_plot}
\alias{distance_plot}
\title{Plot of linkage disequilibrium as a function of the distance between markers}
\usage{
distance_plot(snp_data, center_pos, r2_threshold = NULL)
}
\arguments{
\item{snp_data}{A list describing SNP data and containing at least the
elements \code{LD}, \code{\link{read_hapmap}} for more information on
the contents of this object.}

\item{center_pos}{The central position (in base pairs) of the gene for
which haplotypes are being defined.}

\item{r2_threshold}{A numeric of length one. The r2 value that is considered
as a treshold for selecting marker pairs in a given analysis. Results
in a red horizontal line being added to the graph. \code{NULL} by default,
in which case no horizontal line is added to the graph.}
}
\value{
A graphical object inheriting from classes \code{gg} and \code{ggplot}
  from package \code{ggplot2}, which can be either assigned to an object
  or printed to a graphical device.
}
\description{
This is a high-level plotting function used to plot the linkage disequilibrium
between pairs of markers as a function of the distance between them. One
should expect a pattern of decreasing linkage disequilibrium with increasing
distance. It is important to note that only pairs of markers that are
across the gene are represented, as these are the only ones relevant to
haplotype generation.
}
\details{
As the output of this function is a graphical object from package \code{ggplot2},
it is possible to customize the appearance of the graph and add layers to
the graph using the syntax of this package.
}
\examples{
NULL

}
