% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_plot.R
\name{genotype_plot}
\alias{genotype_plot}
\title{Title}
\usage{
genotype_plot(snp_data, gene_pos = NULL, kept_markers = NULL,
  assignment = NULL, name_order = FALSE)
}
\arguments{
\item{snp_data}{To be completed}

\item{gene_pos}{To be completed}

\item{kept_markers}{To be completed}

\item{assignment}{To be completed}

\item{name_order}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
