% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmgia-dataset.R
\docType{data}
\name{gmgia_haplotypes}
\alias{gmgia_haplotypes}
\title{Analysis of gene \emph{GmGia} of soybean}
\format{To be completed}
\usage{
gmgia_haplotypes
}
\description{
Analysis of gene \emph{GmGia} of soybean
}
\keyword{datasets}
