% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_logfile.R
\name{haplo_logfile}
\alias{haplo_logfile}
\title{Generate the logfile of a haplotype analysis}
\usage{
haplo_logfile(results, filename = "Log.txt", to_file = TRUE)
}
\arguments{
\item{results}{A list containing a variable number of items among the following :
\code{Parameters}, \code{Kinship}, \code{Structure}, \code{All_markers},
\code{Filtered_markers}, \code{Clustered_markers}, \code{Selected_clusters},
\code{Selected_markers}, \code{Haplotypes}. The precise output of the function
will depend on the number of these elements given as input. The function
will throw an error if any element other than those aforementioned is present
in the input list.}

\item{filename}{A character. The name of the file to which the logfile should
be written. Ignored if \code{to_file} is FALSE.}

\item{to_file}{A logical. Should the output be written to file? Defaults to
TRUE, otherwise output is directed to standard output.}
}
\value{
NULL, invisibly. This function is called for its side effect of writing
  to file or to standard output.
}
\description{
This function inspects the output of the \code{\link{haplo_selection}} function
in order to generate a logfile summarizing details of the analysis. The logfile
can be output to a file (defaults to "Log.txt") or to optionally to standard
output.
}
\details{
The logfile is divided into 4 parts :
\itemize{
\item{Analysis parameters} {A description of the input files and parameters that
  were used for the analysis.}
\item{Marker clustering and filtering results} {A description of the number of
  markers that were kept following different filtering steps.}
\item{Selecting marker pairs in LD across gene} {Data regarding the final set
  of markers kept for haplotype generation.}
\item{Haplotype assignment} {A summary of the number of different haplotypes
  generated and the number of individuals unambiguously assigned to each
  haplotype, plus the number of individuals that could not be unambiguously
  assigned to a haplotype due to missing data.}
}
}
\examples{
NULL

}
