% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_params.R
\name{haplo_params}
\alias{haplo_params}
\title{Title}
\usage{
haplo_params(input_file, kinship_file = NULL, structure_file = NULL,
  gene_db_file = NULL, chr_db_file = NULL, gene_name = NULL,
  gene_chrom = NULL, gene_start = NULL, gene_end = NULL,
  chr_length = NULL, file_format = NULL, R2_measure = "r2",
  min_alt_threshold = NULL, max_het_threshold = 0,
  max_missing_threshold = 0, min_allele_count = 1,
  cluster_threshold = 0.95, cluster_R2 = R2_measure,
  marker_independence_threshold = 0.5, max_flanking_pair_distance = 3 *
  10^6, max_marker_to_gene_distance = 10^12)
}
\arguments{
\item{input_file}{To be completed}

\item{kinship_file}{To be completed}

\item{structure_file}{To be completed}

\item{gene_db_file}{To be completed}

\item{chr_db_file}{To be completed}

\item{gene_name}{To be completed}

\item{gene_chrom}{To be completed}

\item{gene_start}{To be completed}

\item{gene_end}{To be completed}

\item{chr_length}{To be completed}

\item{file_format}{To be completed}

\item{R2_measure}{To be completed}

\item{min_alt_threshold}{To be completed}

\item{max_het_threshold}{To be completed}

\item{max_missing_threshold}{To be completed}

\item{min_allele_count}{To be completed}

\item{cluster_threshold}{To be completed}

\item{cluster_R2}{To be completed}

\item{marker_independence_threshold}{To be completed}

\item{max_flanking_pair_distance}{To be completed}

\item{max_marker_to_gene_distance}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
