% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplo_selection.R
\name{haplo_selection}
\alias{haplo_selection}
\title{Generate a selection of markers appropriate for haplotype generation}
\usage{
haplo_selection(analysis_parameters, verbose = TRUE)
}
\arguments{
\item{analysis_parameters}{a list of analysis parameters generated by the function
\code{build.analysis.parameters}.}

\item{verbose}{To be completed.}
}
\value{
A snp_data containing the subset of the chosen markers.
}
\description{
This function applies a series of filters on the genotype data provided (either
in vcf or hapmap format) and outputs a list of markers kept at different stages
of the analysis.
}
\details{
The behaviour of the function can be finely tuned by providing the function call
with an \code{analysis_parameters} object with appropriate contents.

See function \code{genotype.filter} for a list of criteria used to filter the
markers kept for haplotype generation.

So far, the function can only take hapmap and vcf files as input. The name of
the input file must be provided in the list of analysis parameters, but there
should also be an option allowing the user to provide a genotype object as input
to this function. This would allow the user an easier way to look at their input
data before carrying the analysis.
}
\examples{
NULL

}
