% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HaplotypeMiner.R
\docType{package}
\name{HaplotypeMiner}
\alias{HaplotypeMiner}
\alias{HaplotypeMiner-package}
\title{HaplotypeMiner: Gene-centered haplotyping from SNP data}
\description{
This package allows the selection of polymorphic markers useful in defining
haplotypes for genes from SNP data. Included are functions used to parse genotypic
data (VCF and hapmap files), basic tools for filtering SNP data, as well as high-level
data visualization tools.
}
