% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotypes.R
\name{haplotypes}
\alias{haplotypes}
\title{Title}
\usage{
haplotypes(snp_data, kept_markers, center_pos = NULL)
}
\arguments{
\item{snp_data}{To be completed}

\item{kept_markers}{To be completed}

\item{center_pos}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
