% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_valid.R
\name{is_valid}
\alias{is_valid}
\title{Title}
\usage{
is_valid(snp_data, center_pos)
}
\arguments{
\item{snp_data}{To be completed}

\item{center_pos}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
