% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_plot.R
\name{ld_plot}
\alias{ld_plot}
\title{Generates a tile plot of the LD of the given snp_data.}
\usage{
ld_plot(snp_data, center_pos, kept_markers = NULL)
}
\arguments{
\item{snp_data}{The snp_data to be plotted.}

\item{center_pos}{The central position on either side of which markers must be in LD.}

\item{kept_markers}{To be completed.}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
