% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_subset.R
\name{marker_subset}
\alias{marker_subset}
\title{Subsets a snp_data by keeping only certain markers.}
\usage{
marker_subset(snp_data, indices)
}
\arguments{
\item{snp_data}{The snp_data to be subset.}

\item{indices}{Indices of the markers to be kept.}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
