% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2expr.R
\name{r2expr}
\alias{r2expr}
\title{Convert r^2 measure into a printable expression}
\usage{
r2expr(r2_measure)
}
\arguments{
\item{r2_measure}{A character of length one. The R-squared measure used
to represent linkage disequilibrium. Must be either \code{"r2"},
\code{"r2v"}, \code{"r2s"} or \code{"r2vs"}, otherwise an error will
be thrown.}
}
\value{
An object of class \code{expression} which can be used to generate
  pretty labels for graphs.
}
\description{
This function is used internally by package \code{HaplotypeMiner} to generate
pretty labels for graphs.
}
\examples{
NULL

}
