% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hapmap.R
\name{read_hapmap}
\alias{read_hapmap}
\title{Parse a hapmap file.}
\usage{
read_hapmap(filename)
}
\arguments{
\item{filename}{A character. The name of the file to be parsed.}
}
\value{
To be completed.
}
\description{
This function takes a hapmap file as input and converts it to a \code{SnpMatrix}
object.
}
\details{
Details
}
\examples{
NULL

}
