% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kinship.R
\name{read_kinship}
\alias{read_kinship}
\title{Title}
\usage{
read_kinship(kinship_file)
}
\arguments{
\item{kinship_file}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
