% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_structure.R
\name{read_structure}
\alias{read_structure}
\title{Title}
\usage{
read_structure(structure_file)
}
\arguments{
\item{structure_file}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
