% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Parse a vcf file.}
\usage{
read_vcf(filename)
}
\arguments{
\item{filename}{The file to be parsed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
