% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_markers.R
\name{restore_markers}
\alias{restore_markers}
\title{Title}
\usage{
restore_markers(snp_original, snp_target)
}
\arguments{
\item{snp_original}{To be completed}

\item{snp_target}{To be completed}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
