% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subcluster.R
\name{subcluster}
\alias{subcluster}
\title{Drop markers in perfect LD with adjacent markers, keeping only a single
representative marker.}
\usage{
subcluster(snp_data, block_threshold = 1, reverse = FALSE)
}
\arguments{
\item{snp_data}{The snp_data to be filtered.}

\item{block_threshold}{The minimum R^2 for two adjacent markers to be
eligible for blocking (default: 1).}

\item{reverse}{If true, blocks start at the most 3' position.}
}
\value{
A subset of the original \code{snp_data} containing only one
     representative marker per LD block.
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
