% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hapmap.R
\name{write_hapmap}
\alias{write_hapmap}
\title{Export an snp_data into the hapmap format.}
\usage{
write_hapmap(snp_data, filename)
}
\arguments{
\item{snp_data}{The snp_data to be exported.}

\item{filename}{Filename for the output file.}
}
\value{
To be completed
}
\description{
Description
}
\details{
Details
}
\examples{
NULL

}
