% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_vcf.R
\name{write_vcf}
\alias{write_vcf}
\title{Write VCF output}
\usage{
write_vcf(snp_data, filename)
}
\arguments{
\item{snp_data}{The object to be written to the file.}

\item{filename}{The name of the output file.}
}
\value{
To be completed
}
\description{
This will generate a VCF file containing the genotypes. This will only work if the
initial file was in VCF format.
}
\details{
Details
}
\examples{
NULL

}
